refactor: 패키지 구조 변경
line, station 패키지로 나뉘어진 상태에서 좀 더 세분화하여 구분하였습니다.
흔히 많이 사용하는 구조인 controller, service, repository 에 dto 패키지를 추가하여 구성하였습니다.
좋은 구조라고 하기에는 어려울 수 있지만, 간단한 구조에서 빠른 개발을 함에 있어서는 좋은 구조라고 생각합니다.
