--- src/main/java/subway/line/LineRepository.java
@@ -1,6 +1,0 @@
-package subway.line;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface LineRepository extends JpaRepository<Line, Long> {
-}
