+++ src/main/java/subway/line/dto/UpdateLineRequest.java
@@ -1,8 +1,9 @@
-package subway.line;
+package subway.line.dto;
 
 import lombok.Value;
 
 @Value
+public
 class UpdateLineRequest {
     String name;
     String color;
