+++ src/main/java/subway/line/repository/LineRepository.java
@@ -1,0 +1,6 @@
+package subway.line.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LineRepository extends JpaRepository<Line, Long> {
+}
